This commit introduces a Java utility library
The only utility it contains now is a workaround for
the lack of "lazy" final field initializers in Java.
It makes it more convenient to have "almost final"
semantics on fields, that is to say keep having
final fields everywhere, even if a value needs to
be computed lazily.

There are currently two implementations of lazy
providers, depending on thread-safety. Initial
benchmarks tend to show that the unsafe version
is 4 time faster than the locking one.